Remove the failing test
